requirements.txt
certifi==2023.11.17
chardet==5.2.0
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.13.1
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.1.0
python-barcode==0.15.1
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.8
six==1.16.0
tzdata==2023.4
virtualenv==20.25.0
